requirements.txt
langgraph>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langsmith>=0.2.0
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.24.0
matplotlib>=3.7.0
ccxt>=4.5.19
dydx-v4-client>=1.1.0
protobuf==5.29.5
Pillow>=10.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
